Rank suggestions by sentence-embedding similarity
The old set-overlap score ignored leisure_time and wants_in_relationship
entirely and treated "hiking" and "trekking" as unrelated. Move to a 384-dim
embedding per user via all-MiniLM-L6-v2 (local, free, Apache-2.0) and rank
candidates by pgvector cosine similarity inside a SECURITY DEFINER RPC.

Hard filters (target_intent, term_length) still gate the candidate pool;
embeddings only re-rank within it. Embeddings are computed lazily and
persisted on the row so /suggestions stays a single SQL kNN at read time.

+++ backend/models.py
@@ -115,4 +115,4 @@ class SuggestionResponse(BaseModel):
     avatar_url: Optional[str] = None
     gender: Optional[str] = None
     zodiac_sign: Optional[str] = None
-    score: int = 0
+    score: float = 0.0
